compile warnings

+++ vmgc/src/gcstate.cpp
@@ -127,14 +127,14 @@ namespace vmgc {
 		auto len = _empty_big_buffers->size();
 		if (len > 0) {
 			const auto &frontBuf = _empty_big_buffers->front();
-			if (size <= _empty_big_buffers->front().second) {
+			if (ptrdiff_t(size) <= _empty_big_buffers->front().second) {
 				auto bufpos = frontBuf.first;
 				auto bufsz = frontBuf.second;
 				b.pos = bufpos;
 
 				_empty_big_buffers->pop_front();
 
-				if (size < bufsz) {
+				if (size < size_t(bufsz)) {
 					std::pair<intptr_t, ptrdiff_t> eb;
 					eb.first = bufpos + size;
 					eb.second = bufsz - size;
@@ -147,14 +147,14 @@ namespace vmgc {
 			}
 
 			const auto &backBuf = _empty_big_buffers->back();
-			if (size <= backBuf.second) {
+			if (ptrdiff_t(size) <= backBuf.second) {
 				auto bufpos = backBuf.first;
 				auto bufsz = backBuf.second;
 				b.pos = bufpos;
 
 				_empty_big_buffers->pop_back();
 
-				if (size < bufsz) {
+				if (ptrdiff_t(size) < bufsz) {
 					std::pair<intptr_t, ptrdiff_t> eb;
 					eb.first = bufpos + size;
 					eb.second = bufsz - size;
@@ -172,7 +172,7 @@ namespace vmgc {
 		if (size > DEFAULT_GC_BLOCK_SIZE) {
 			mallocSize = size;
 		}
-		if (mallocSize + _total_malloced_blocks_size > _max_gc_size) {
+		if (ptrdiff_t(mallocSize) + _total_malloced_blocks_size > _max_gc_size) {
 			_used_size -= size;
 			return nullptr;
 		}
